Depecrate export and rename "model" to "train"
