+++ nilm_thresholding/train_model.py
@@ -3,109 +3,10 @@ import warnings
 
 warnings.simplefilter(action="ignore", category=FutureWarning)
 
-import os
-import time
-
 import typer
 
-from nilm_thresholding.data.loader import DataLoader
-from nilm_thresholding.model.model import initialize_model
-from nilm_thresholding.results.store_output import (
-    generate_path_output,
-    generate_folder_name,
-    list_scores,
-    store_scores,
-)
+from nilm_thresholding.model.train import train_many_models
 from nilm_thresholding.utils.config import load_config
-from nilm_thresholding.utils.format_list import merge_dict_list
-
-
-def train_many_models(path_data, path_output, config):
-    """
-    Runs several models with the same conditions.
-    Stores plots and the average scores of those models.
-    """
-    # Set output path
-    path_output = generate_path_output(path_output, config["name"])
-    path_output_folder = generate_folder_name(
-        path_output,
-        config["input_len"],
-        config["period"],
-        config["classification_w"],
-        config["regression_w"],
-        config["threshold"]["method"],
-    )
-
-    # Load dataloader
-    dataloader_train = DataLoader(path_data, subset="train", shuffle=True, **config)
-    dataloader_validation = DataLoader(
-        path_data, subset="validation", shuffle=True, **config
-    )
-    dataloader_test = DataLoader(path_data, subset="test", shuffle=False, **config)
-
-    # Training
-
-    act_scores = []
-    pow_scores = []
-    time_ellapsed = 0
-
-    for i in range(config["num_models"]):
-        print(f"\nModel {i + 1}\n")
-
-        model = initialize_model(config)
-
-        # Train
-        time_start = time.time()
-        model.train(
-            dataloader_train,
-            dataloader_validation,
-        )
-        time_ellapsed += time.time() - time_start
-
-        # Store the model
-        path_model = os.path.join(path_output_folder, f"model_{i}.pth")
-        model.save(path_model)
-
-        act_scr, pow_scr = model.score(dataloader_test)
-
-        act_scores += act_scr
-        pow_scores += pow_scr
-
-        # Store individual scores
-        act_dict = merge_dict_list(act_scr)
-        pow_dict = merge_dict_list(pow_scr)
-
-        scores = {"classification": act_dict, "regression": pow_dict}
-
-        filename = f"scores_{i}.txt"
-
-        store_scores(
-            path_output_folder,
-            config,
-            scores,
-            time_ellapsed,
-            filename=filename,
-        )
-
-    # List scores
-
-    scores = list_scores(
-        config["appliances"],
-        act_scores,
-        pow_scores,
-        config["num_models"],
-    )
-
-    time_ellapsed /= config["num_models"]
-
-    # Store scores and plot
-
-    store_scores(
-        path_output_folder,
-        config,
-        scores,
-        time_ellapsed,
-    )
 
 
 def main(
